Split logging and error functions into new classes in different modules to simplify the code. Helps with further generalisation.
